Move md5sum to utils.py
Move md5sum to utils.py since it makes more sense to keep it there.
