+++ pbench/server/__init__.py
@@ -4,12 +4,10 @@ Simple module level convenience functions.
 
 import os
 import logging
-import hashlib
 
 from logging import handlers
 from time import time as _time
 from datetime import datetime
-from functools import partial
 from configparser import ConfigParser, NoSectionError, NoOptionError
 
 from pbench.lib import configtools
@@ -286,15 +284,3 @@ class PbenchConfig(object):
                   <YYYY>-<MM>-<DD>T<hh>:<mm>:<ss>-<TZ>
         """
         return utils.tstos(_time())
-
-
-def md5sum(filename):
-    """
-    Return the MD5 check-sum of a given file.
-    We don't want to read the entire file into memory.
-    """
-    with open(filename, mode="rb") as f:
-        d = hashlib.md5()
-        for buf in iter(partial(f.read, 128), b""):
-            d.update(buf)
-    return d.hexdigest()
